backend/requirements.txt
fastapi==0.110.3
uvicorn[standard]==0.29.0
requests==2.32.3
requests-cache==1.1.1
pandas==2.2.2
numpy==1.26.4
tenacity==8.2.3
python-dotenv==1.0.1
pydantic==1.10.14
openai==1.30.1
reportlab==4.2.0
plotly==5.22.0
kaleido==0.2.1
scikit-learn==1.4.2
beautifulsoup4==4.12.3
lxml==5.2.1
neo4j==5.18.0
python-dateutil==2.9.0.post0
pyarrow==16.1.0
